removing data renaming app js for now
