--- assignment/assignment02/data/WiredTiger
@@ -1,2 +1,0 @@
-WiredTiger
-WiredTiger 11.3.0: (November 16, 2023)
